requirements.txt
typer
inquirer
yaspin
bcrypt
psycopg2
python-dotenv
pyperclip